+++ packages/pancake-sass/src/sass.js
@@ -41,12 +41,18 @@ import { Log, Style, WriteFile } from '@gov.au/pancake';
 export const GetPath = ( module, modules, baseLocation, npmOrg ) => {
 	let modulePath = '';
 
+	const npmOrgs = npmOrg.split( ' ' );
+	let location;
+	npmOrgs.forEach( org => {
+		if( baseLocation.includes( org ) ){
+			location = baseLocation.replace( `${ org }/`, '' );
+		}
+	});
+
 	for( const item of modules ) {
 		if( item.name === module ) {
-			const moduleName = module.replace(`${ npmOrg }/`, '');
-
 			if( item.pancake['pancake-module'].sass.path ) {
-				modulePath = Path.normalize(`${ baseLocation }/${ moduleName }/${ item.pancake['pancake-module'].sass.path }`);
+				modulePath = Path.normalize(`${ location }/${ module }/${ item.pancake['pancake-module'].sass.path }`);
 			}
 			else {
 				modulePath = false;
@@ -122,7 +128,7 @@ export const GenerateSass = ( location, name, modules, npmOrg ) => {
 
 	if( dependencies ) {
 		for( const dependency of Object.keys( dependencies ) ) {
-			const modulePath = GetPath( dependency, modules, baseLocation, npmOrg )
+			const modulePath = GetPath( dependency, modules, baseLocation, npmOrg );
 
 			if( modulePath ) {
 				sass += `@import "${ modulePath }";\n`;
@@ -130,7 +136,7 @@ export const GenerateSass = ( location, name, modules, npmOrg ) => {
 		}
 	}
 
-	const modulePath = GetPath( name, modules, baseLocation, npmOrg )
+	const modulePath = GetPath( name, modules, baseLocation, npmOrg );
 	sass += `@import "${ modulePath }";\n`;
 
 	return sass.replace(/\\/g, "\\\\"); // escape path for silly windows